Apache Flink/requirements.txt
apache-flink==1.18.0
kafka-python==2.0.2
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.34.0
requests==2.31.0
pydantic==2.5.0